[Keyboard] Move hhkb_yang to hhkb/yang
